Add files via upload
Some work I did for my business communications class based on dummy data.  To clarify, this is not actual documentation or use of actual data from State Farm (letterhead was a requirement from the course and was grabbed from the website).  All data has been adjusted accordingly to simulate the process of obtaining real findings.